Extract list/grid manga card
